refactor from reviews
